requirement.txt
Django==4.2.7
django-cors-headers==4.3.1
django-crontab==0.7.1
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dotenv==1.0.0
razorpay==1.4.1
django-tinymce
pymysql
jsons
cryptography
